refactor(product): Hot 캐시 TTL 값을 60분에서 30분으로 변경
- Hot 캐시의 TTL을 30분으로 조정하여 데이터 갱신 주기를 단축
- 관련 주석 및 문서 업데이트
- MV 테이블 동기화 및 캐시 갱신 로직 개선
