+++ apps/commerce-api/src/main/java/com/loopers/infrastructure/product/ProductMVBatchScheduler.java
@@ -1,11 +1,22 @@
 package com.loopers.infrastructure.product;
 
+import java.util.List;
+import java.util.stream.Collectors;
+
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
 import com.loopers.application.product.BatchUpdateResult;
+import com.loopers.domain.brand.BrandEntity;
+import com.loopers.domain.brand.BrandService;
 import com.loopers.domain.product.ProductCacheService;
+import com.loopers.domain.product.ProductMaterializedViewEntity;
 import com.loopers.domain.product.ProductMVService;
+import com.loopers.infrastructure.cache.CacheStrategy;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -13,10 +24,11 @@ import lombok.extern.slf4j.Slf4j;
 /**
  * ProductMaterializedView 배치 스케줄러
  *
- * <p>2분 간격으로 MV 테이블을 동기화하고, 20분마다 Hot 캐시를 갱신합니다.</p>
- *
- * @author hyunjikoh
- * @since 2025. 11. 10.
+ * <p>MV 테이블 동기화 및 Hot 캐시 갱신을 통합 관리합니다.</p>
+ * <ul>
+ *   <li>2분마다: MV 테이블 동기화</li>
+ *   <li>50분마다: Hot 캐시 갱신 (인기순 상품, 브랜드별 인기순)</li>
+ * </ul>
  */
 @Component
 @RequiredArgsConstructor
@@ -24,16 +36,19 @@ import lombok.extern.slf4j.Slf4j;
 public class ProductMVBatchScheduler {
     private final ProductMVService mvService;
     private final ProductCacheService cacheService;
+    private final BrandService brandService;
+
+    // 배치 갱신 설정
+    private static final int TOP_PRODUCTS_COUNT = 100;
+    private static final int CACHE_PAGES_PER_BRAND = 3;
+    private static final int PAGE_SIZE = 20;
 
     /**
-     * MV 테이블 동기화 배치 작업
-     *
-     * <p>2분(120초) 간격으로 실행되며, 변경된 데이터를 MV 테이블에 반영합니다.</p>
-     * <p>변경된 상품과 영향받은 브랜드의 캐시만 선택적으로 무효화합니다.</p>
+     * MV 테이블 동기화 배치 작업 (2분마다)
      */
-    @Scheduled(fixedDelay = 120000) // 2분 간격
+    @Scheduled(fixedDelay = 120000)
     public void syncMaterializedView() {
-        log.info("=== MV 배치 업데이트 시작 ===");
+        log.info("MV 배치 업데이트 시작");
 
         try {
             BatchUpdateResult result = mvService.syncMaterializedView();
@@ -43,46 +58,134 @@ public class ProductMVBatchScheduler {
                 return;
             }
 
-            // 변경사항이 있는 경우 Incremental 캐시 무효화
             if (result.hasChanges()) {
-                log.info("변경사항 감지 - 선택적 캐시 무효화 시작");
-                log.info("- 변경된 상품: {}개", result.getChangedProductIds().size());
-                log.info("- 영향받은 브랜드: {}개", result.getAffectedBrandIds().size());
-
-                // 변경된 상품과 브랜드의 캐시만 무효화
+                log.info("변경사항 감지 - 선택적 캐시 무효화");
                 cacheService.evictCachesAfterMVSync(
                         result.getChangedProductIds(),
                         result.getAffectedBrandIds()
                 );
             } else {
-                log.info("변경사항 없음 - 캐시 무효화 건너뜀");
+                log.info("변경사항 없음");
             }
 
-            log.info("=== MV 배치 업데이트 완료 === 생성: {}건, 갱신: {}건, 소요: {}ms",
+            log.info("MV 배치 업데이트 완료 - 생성: {}건, 갱신: {}건, 소요: {}ms",
                     result.getCreatedCount(), result.getUpdatedCount(), result.getDurationMs());
 
         } catch (Exception e) {
-            log.error("MV 배치 업데이트 중 예상하지 못한 오류 발생", e);
+            log.error("MV 배치 업데이트 중 오류", e);
         }
     }
 
     /**
-     * Hot 캐시 갱신 배치 작업
+     * Hot 캐시 갱신 배치 작업 (50분마다)
      *
-     * <p>20분마다 실행되며, 캐시 스탬피드를 방지하기 위해 TTL 만료 전에 캐시를 갱신합니다.</p>
+     * <p>배치 갱신으로 캐시 스탬피드 방지</p>
+     * <p>ProductMVRepository를 직접 사용하여 likeCount 정렬 보장</p>
      */
-    @Scheduled(cron = "0 */20 * * * *") // 20분마다
+    @Scheduled(fixedRate = 50 * 60 * 1000, initialDelay = 60 * 1000)
     public void refreshHotCache() {
         log.info("Hot 캐시 갱신 시작");
 
-        try {
-            // TODO: Hot 데이터 캐시 갱신 로직 구현
-            // - 전체 상품 목록 첫 2페이지 캐시 갱신
-            // - 각 브랜드별 상품 목록 첫 2페이지 캐시 갱신
+        long startTime = System.currentTimeMillis();
 
-            log.info("Hot 캐시 갱신 완료");
+        try {
+            refreshPopularProductDetails();
+            refreshBrandPopularProductIds();
+
+            long duration = System.currentTimeMillis() - startTime;
+            log.info("Hot 캐시 갱신 완료 - {}ms", duration);
+
         } catch (Exception e) {
             log.error("Hot 캐시 갱신 실패", e);
         }
+    }
+
+    /**
+     * 인기 상품 상세 정보 갱신 (likeCount 상위 100개)
+     */
+    private void refreshPopularProductDetails() {
+        log.debug("인기 상품 상세 갱신");
+
+        try {
+            // ✅ ProductMVRepository 직접 사용 (likeCount 정렬 보장)
+            Pageable pageable = PageRequest.of(0, TOP_PRODUCTS_COUNT,
+                    Sort.by(Sort.Direction.DESC, "likeCount"));
+
+            Page<ProductMaterializedViewEntity> popularProducts = mvService.findAll(pageable);
+
+            List<Long> productIds = popularProducts.getContent().stream()
+                    .map(ProductMaterializedViewEntity::getProductId)
+                    .collect(Collectors.toList());
+
+            cacheService.cacheProductIds(CacheStrategy.HOT, null, pageable, productIds);
+
+            log.info("인기 상품 상세 갱신 완료 - {}개", productIds.size());
+
+        } catch (Exception e) {
+            log.error("인기 상품 상세 갱신 실패", e);
+        }
+    }
+
+    /**
+     * 브랜드별 인기순 상품 ID 리스트 갱신 (첫 3페이지)
+     */
+    private void refreshBrandPopularProductIds() {
+        log.debug("브랜드별 인기순 ID 리스트 갱신");
+
+        try {
+            List<BrandEntity> brands = brandService.getAllBrands();
+            int totalRefreshed = 0;
+
+            for (BrandEntity brand : brands) {
+                try {
+                    int refreshed = refreshBrandProductIds(brand.getId());
+                    totalRefreshed += refreshed;
+                } catch (Exception e) {
+                    log.warn("브랜드 ID 리스트 갱신 실패 - brandId: {}", brand.getId());
+                }
+            }
+
+            log.info("브랜드별 ID 리스트 갱신 완료 - {}개 브랜드, {}페이지",
+                    brands.size(), totalRefreshed);
+
+        } catch (Exception e) {
+            log.error("브랜드별 ID 리스트 갱신 실패", e);
+        }
+    }
+
+    /**
+     * 특정 브랜드의 인기순 상품 ID 리스트 갱신
+     */
+    private int refreshBrandProductIds(Long brandId) {
+        int refreshedPages = 0;
+        Sort sort = Sort.by(Sort.Direction.DESC, "likeCount");
+
+        for (int page = 0; page < CACHE_PAGES_PER_BRAND; page++) {
+            try {
+                Pageable pageable = PageRequest.of(page, PAGE_SIZE, sort);
+
+                // ✅ ProductMVRepository 직접 사용 (likeCount 정렬 보장)
+                Page<ProductMaterializedViewEntity> products =
+                        mvService.findByBrandId(brandId, pageable);
+
+                if (products.isEmpty()) {
+                    break;
+                }
+
+                List<Long> productIds = products.getContent().stream()
+                        .map(ProductMaterializedViewEntity::getProductId)
+                        .collect(Collectors.toList());
+
+                cacheService.cacheProductIds(CacheStrategy.HOT, brandId, pageable, productIds);
+                refreshedPages++;
+
+                log.debug("브랜드 ID 리스트 갱신 - brandId: {}, page: {}", brandId, page);
+
+            } catch (Exception e) {
+                log.warn("브랜드 ID 리스트 갱신 실패 - brandId: {}, page: {}", brandId, page);
+            }
+        }
+
+        return refreshedPages;
     }
 }
